#59 feat: api 도메인에 따라 users 파일명으로 queries 통일
